[Bugfix] UNSHIFT sends Shift state and key simultaneously.
The previous implementation unregisters/registers the Shift modifier
first and then registers/unregisters the key. It generates an
intermediate state that only Shift state is changed. The unshifted
result of the key held prior to UNSHIFT can be observed during this
intermediate state. This intermediate state is eliminated by this fix.

Let me show the problematic scenario. The key 2 is an UNSHIFT key.

1. Turn on US_ON_JIS layer
2. Hold Left Shift
3. Hold J -- sequence of capital J
4. Hold 2 -- sequence of left square bracket

The result should be something like `JJJJJJ[[[[[[[` if the system
key layout is US, or `JJJJJJJ@@@@@@ if that was JIS. Let me assume
system layout is US. I occasinally observed something like
`JJJJJj[[[[[`. The report sequence sent was as follows:

    keyboard_report: 02 00 00 00 00 00 00 00 // hold LShift
    keyboard_report: 02 00 00 00 00 00 00 00
    keyboard_report: 02 00 0D 00 00 00 00 00 // hold J
    keyboard_report: 00 00 0D 00 00 00 00 00 // hold 2 -- (#)
    keyboard_report: 00 00 0D 2F 00 00 00 00
    keyboard_report: 00 00 0D 00 00 00 00 00 // release 2
    keyboard_report: 02 00 0D 00 00 00 00 00
    keyboard_report: 02 00 00 00 00 00 00 00 // release J
    keyboard_report: 02 00 00 00 00 00 00 00
    keyboard_report: 00 00 00 00 00 00 00 00 // release LShift
    keyboard_report: 00 00 00 00 00 00 00 00

Even if the appeared result was as expected `JJJJJ[[[[[` or unexpected
`JJJJJj[[[[[`, the report was the same. I guess that depends on timing.
Anyway, OS had a chance to generate small `j` when handling report (#).
The fixed report sequence is as follows:

    keyboard_report: 02 00 00 00 00 00 00 00 // hold LShift
    keyboard_report: 02 00 00 00 00 00 00 00
    keyboard_report: 02 00 0D 00 00 00 00 00 // hold J
    keyboard_report: 00 00 0D 2F 00 00 00 00 // hold 2
    keyboard_report: 02 00 0D 00 00 00 00 00 // release 2
    keyboard_report: 02 00 00 00 00 00 00 00 // release J
    keyboard_report: 02 00 00 00 00 00 00 00
    keyboard_report: 00 00 00 00 00 00 00 00 // release LShift
    keyboard_report: 00 00 00 00 00 00 00 00

There is no more chance to generate small letter `j` because there is no
report composed of only J.

As it is easily imagined, there is the counter part of this behaviour.

1. Turn on US_ON_JIS layer
2. Hold Left Shift
3. Hold 2 -- sequence of left square bracket
4. Hold J -- sequence of capital J

We occasinally observe a left curly brace between the sequence of left
square brackets and capital J along to this scenario because the report
sequence sent is as follows:

    keyboard_report: 02 00 00 00 00 00 00 00 // hold LShift
    keyboard_report: 00 00 2F 00 00 00 00 00 // hold 2
    keyboard_report: 02 00 2F 00 00 00 00 00 // hold J --- (*)
    keyboard_report: 02 00 2F 0D 00 00 00 00
    keyboard_report: 02 00 2F 00 00 00 00 00 // release J
    keyboard_report: 00 00 2F 00 00 00 00 00
    keyboard_report: 02 00 00 00 00 00 00 00 // release 2
    keyboard_report: 00 00 00 00 00 00 00 00 // release LShift

Press and release of `S(KC_J)` sends the report twice. There is a chance
of left curly brace to appear on (*). Fix of the function
`process_action` defined in `tmk_core/common/action.c` is required to
fix this behaviour. Or, we can use one FN key dedicated to the shifted
normal keys in SHIFT_US_ON_JIS layer.

I took a glance of `process_action` and could not find out the reason to
send a report that handles only modifier change before the report for
the whole action.