C-H is added as a rigid keybind. It works as Backspace.
It is really interesting that we have to make uncontrol_count variable
be a global one. We can avoid that in 2 strategy I adopted before:

  1. Use only one Layer and one Function key
    - FIXED_RIGHT_PINKY_ON_ASCII and Fn4 is the example.
  2. Use 2 Layers and 2 Function keys but control them in one function
    - FIXED_RIGHT_PINKY_ON_JIS and SHIFTED_FIXED_RIGHT_PINKY_ON_JIS
    - Fn5 and Fn3

FIXED_RIGHT_PINKY_ON_ASCII approach does not only saves the number of
layers and function keys consumed but also allows us to control that
feature by only one function (or, say, a function closure).