Number keys except 0 and 1 are shifted 1 key right.
The concept of this modification is the use of fingers. First, I
personally my ring finger for 1 and 2, middle finger for 3, index finger
for 4 and 5. I want to use same fingers as much as possible and the
first row is decided based on that concept. The modification done in
this comimt fixes the correspondence on the third row.

0 and 1 are decided by stronger requirements: easy to use. These are
mathematically very important numbers (... as numbers not characters,
though) and also very important for me because I use 0 so much on VIM
and 1 on Tig. Even on the programming, these are used frequently
compeared to other digits.

In addition, I use my ring finger to issue `!` since I always use left
shift.