Right hand side first row of Symbols Layers are moved to left.
I want to avoid to  change the use of fingers as much as possible. The
new first row issues,

  + `&` and `*` by the right index finger
  + `_` by the right middle finger
  + `0` by the right ring finger
  + `]` by the right pinky

and all of them are same as the use of fingers in normal size keyboards.
Initially, I located `7` on the Numbers layer to make it easy to push.
Even though the new position is a bit harder than before, I can handle
it.